Unlock Player's names for buttons. Avoid high privileges purge methods on channel. Enable modularity in discord_bot and move unused code to modues
